refactor: add AdminAccess page to the "admin-backpanel" module and update import path in routes
